requirements.txt
crewai>=0.80.0
crewai-tools>=0.14.0
langchain-openai>=0.2.8
python-dotenv>=1.0.1
